Migrate public-key to 'api' module
