+++ solr/core/src/java/org/apache/solr/security/GetPublicKey.java
@@ -17,45 +17,31 @@
 
 package org.apache.solr.security;
 
-import static org.apache.solr.client.solrj.impl.BinaryResponseParser.BINARY_CONTENT_TYPE_V2;
-
-import com.fasterxml.jackson.annotation.JsonProperty;
-import io.swagger.v3.oas.annotations.media.Schema;
 import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
 import org.apache.solr.api.JerseyResource;
-import org.apache.solr.client.api.model.SolrJerseyResponse;
+import org.apache.solr.client.api.endpoint.GetPublicKeyApi;
+import org.apache.solr.client.api.model.PublicKeyResponse;
 import org.apache.solr.jersey.PermissionName;
 
 /**
- * V2 API for fetching the public key of the receiving node.
+ * V2 API implementation to fetch the public key of the receiving node.
  *
  * <p>This API is analogous to the v1 /admin/info/key endpoint.
  */
-@Path("/node/key")
-public class PublicKeyAPI extends JerseyResource {
+public class GetPublicKey extends JerseyResource implements GetPublicKeyApi {
 
   private final SolrNodeKeyPair nodeKeyPair;
 
   @Inject
-  public PublicKeyAPI(SolrNodeKeyPair nodeKeyPair) {
+  public GetPublicKey(SolrNodeKeyPair nodeKeyPair) {
     this.nodeKeyPair = nodeKeyPair;
   }
 
-  @GET
-  @Produces({"application/json", "application/xml", BINARY_CONTENT_TYPE_V2})
+  @Override
   @PermissionName(PermissionNameProvider.Name.ALL)
   public PublicKeyResponse getPublicKey() {
     final PublicKeyResponse response = instantiateJerseyResponse(PublicKeyResponse.class);
     response.key = nodeKeyPair.getKeyPair().getPublicKeyStr();
     return response;
-  }
-
-  public static class PublicKeyResponse extends SolrJerseyResponse {
-    @JsonProperty("key")
-    @Schema(description = "The public key of the receiving Solr node.")
-    public String key;
   }
 }
